"Update info icon availability and card flip behavior"
This commit introduces two changes in the flashcard's behavior. First, the info icon will now only appear if the phrase has a definition, thus not misleading the user. Previously, it was always present leading to potential confusion if there was no definition available.

Secondly, we updated the card flip behavior. Now, the card only flips when it is clicked, but not when the info icon or the popup from the info icon is clicked. Before, the card additionally flipped when these elements were clicked, which was not the desired behavior.

The styles were also adjusted to match the new behavior. The hover state was replaced by a class, to be added and removed via JavaScript.