Fix flipCard function in flashcard.js
Modified the conditional statement in the flipCard function to remove duplicated code and correctly implemented the flip action. Previously, the 'flipped' class was toggled on 'this' regardless of the target. Now, 'flipped' class is toggled directly on 'currentTarget' which is the actual card that should flip, ensuring the function works as expected.