Your commit message is incorrect according to the given diff. Here is the correction:
Add new phrases related to 'dwelling on' in phrases.json

This commit introduces three new phrases in Spanish and their respective English translations to the phrases.json file. All phrases illustrate different contexts and uses of the English expression "dwelling on". The update aims to enhance the language learning resource with more diverse and practical phrases.