Add new phrase explaining the use of the preposition 'for'
A phrase demonstrating the use of the preposition 'for' has been introduced to the 'phrases.json' file. The phrase "¿Cuánto tiempo te llevó convertirte en arquitecto?" in Spanish translates to "How long did it take for you to become an architect?" in English. This phrase illustrates the use of 'for' to indicate the duration of time to complete a process, in this case, becoming an architect.