Add new Spanish-English phrases to phrases.json
New phrases and their corresponding translations have been added to the 'phrases.json' file. These additions included phrases with the following English translations: "It was nice getting to know you a bit more.", "Hang on, could you pass me the salt?", "Hang on, could you check if you’re free tomorrow?", "Hang on, could you help me with my computer?", and "Hang on, could you explain that to me again?". The aim of this commit is to enrich the database with more diverse phrases to enhance the translation model's accuracy and versatility.