Improve phrase processing in flashcard.js
The processPhrases method has been updated to shuffle the phrases more thoroughly before they are passed to the Card constructor. A standard Fisher-Yates shuffle algorithm is implemented to randomize the order of phrases, replacing the previous sort method. This approach should ensure that the flashcards are presented in a truly random order.