The provided commit message doesn't match with the code differences given. Here is the appropriate commit message based on the code differences.
Commit message:
Rename card creation methods in flashcard.js

The methods for creating the front and back of the flashcard in flashcard.js have been renamed from createCardFront and createCardBack to createFrontSide and createBackSide. Corresponding changes have been made in the function where these methods are called ( in divChild.appendChild). These changes aim to clarify the purpose of the methods.